Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
requests
psycopg2
python-dotenv